fix (core,http-client): Don't parse failed HTTP responses
