+++ packages/common/package.json
@@ -34,6 +34,7 @@
     "@ceramicnetwork/streamid": "^1.0.0",
     "@overnightjs/logger": "^1.2.0",
     "cids": "~1.1.6",
+    "cross-fetch": "^3.1.4",
     "flat": "^5.0.2",
     "lodash.clonedeep": "^4.5.0",
     "logfmt": "^1.3.2",
