+++ packages/common/src/index.ts
@@ -3,6 +3,7 @@ export * from './ceramic-api'
 export * from './context'
 export * from './docopts'
 export * from './stream'
+export * from './utils/http-utils'
 export * from './utils/stream-utils'
 export * from './utils/test-utils'
 export * from './logger-provider'
