+++ packages/core/package.json
@@ -45,7 +45,6 @@
     "ajv-formats": "^2.0.2",
     "await-semaphore": "^0.1.3",
     "cids": "~1.1.6",
-    "cross-fetch": "^3.1.4",
     "dids": "^2.1.0",
     "ipld-dag-cbor": "^0.17.0",
     "level-ts": "^2.0.5",
