+++ packages/http-client/package.json
@@ -31,7 +31,6 @@
     "@ceramicnetwork/stream-caip10-link": "^1.0.1",
     "@ceramicnetwork/stream-tile": "^1.0.1",
     "@ceramicnetwork/streamid": "^1.0.0",
-    "cross-fetch": "^3.1.4",
     "query-string": "7.0.0",
     "rxjs": "^7.0.0"
   },
