+++ packages/http-client/src/utils.ts
@@ -1,24 +1,4 @@
-import fetch from "cross-fetch"
 import StreamID from "@ceramicnetwork/streamid"
-
-interface FetchOpts {
-  body?: any
-  method?: string
-}
-
-export async function fetchJson(url: string, opts: FetchOpts = {}): Promise<any> {
-    if (opts.body) {
-        Object.assign(opts, {
-            body: JSON.stringify(opts.body),
-            headers: { 'Content-Type': 'application/json' }
-        })
-    }
-    const res = await fetch(url, opts).then(response => response.json())
-    if (res.error) {
-        throw new Error(res.error)
-    }
-    return res
-}
 
 export function typeStreamID(streamId: StreamID | string): StreamID  {
     return (typeof streamId === 'string') ? StreamID.fromString(streamId) : streamId
